Fixed CoordinatorLayout bug
So it seems that the CoordinatorLayout, because it's a FrameLayout,
which is traditionally used to stack views on top of one another,
doesn't do a very good job of arranging views linearly and assigning
them with proper heights. Hence the ViewPager was too tall for the
screen making part of it extend below the bottom edge of the activity
window. To solve this I wrapped the AppBarLayout and the ViewPager in a
LinearLayout nested inside the CoordinatorLayout.